🚚 (mon-pix) move module tests
