+++ mon-pix/tests/unit/routes/module-test.js
@@ -5,6 +5,31 @@ import sinon from 'sinon';
 module('Unit | Route | module', function (hooks) {
   setupTest(hooks);
 
+  test('should exist', function (assert) {
+    // when
+    const route = this.owner.lookup('route:module');
+
+    // then
+    assert.ok(route);
+  });
+
+  test('should load the corresponding model', async function (assert) {
+    // given
+    const route = this.owner.lookup('route:module');
+    const store = this.owner.lookup('service:store');
+
+    const module = Symbol('the-module');
+
+    store.queryRecord = sinon.stub();
+    store.queryRecord.withArgs('module', { shortId: 'dfc4isne', encryptedRedirectionUrl: 'somehash' }).resolves(module);
+
+    // when
+    const model = await route.model({ shortId: 'dfc4isne', redirection: 'somehash' });
+
+    // then
+    assert.strictEqual(model, module);
+  });
+
   test('activate should set metrics context', function (assert) {
     // Given
     const route = this.owner.lookup('route:module');
