--- mon-pix/tests/unit/routes/module/module-test.js
@@ -1,32 +1,0 @@
-import { setupTest } from 'ember-qunit';
-import { module, test } from 'qunit';
-import sinon from 'sinon';
-
-module('Unit | Route | modules | module', function (hooks) {
-  setupTest(hooks);
-
-  test('should exist', function (assert) {
-    // when
-    const route = this.owner.lookup('route:module');
-
-    // then
-    assert.ok(route);
-  });
-
-  test('should load the corresponding model', async function (assert) {
-    // given
-    const route = this.owner.lookup('route:module');
-    const store = this.owner.lookup('service:store');
-
-    const module = Symbol('the-module');
-
-    store.queryRecord = sinon.stub();
-    store.queryRecord.withArgs('module', { shortId: 'dfc4isne', encryptedRedirectionUrl: 'somehash' }).resolves(module);
-
-    // when
-    const model = await route.model({ shortId: 'dfc4isne', redirection: 'somehash' });
-
-    // then
-    assert.strictEqual(model, module);
-  });
-});
